frontend-cal/backend-cal/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
pillow
requests
pydantic